+++ tests/e2e/matchers/index.js
@@ -1,0 +1,1 @@
+export { toHaveAdsenseTag } from './to-have-adsense-tag';
